# Optimize methods of Klass for performance.