--- a/lib/rails_best_practices/core/klasses.rb
+++ b/lib/rails_best_practices/core/klasses.rb
@@ -14,19 +14,14 @@ module RailsBestPractices
 
     # Class info includes class name, extend class name and module names.
     class Klass
+      attr_reader :extend_class_name, :class_name
+
       def initialize(class_name, extend_class_name, modules)
-        @class_name = class_name
-        @extend_class_name = extend_class_name
         @modules = modules.dup
-      end
-
-      def class_name
-        @modules.map { |modu| "#{modu}::" }.join("") + @class_name
-      end
-
-      def extend_class_name
-        if @extend_class_name.present?
-          @modules.map { |modu| "#{modu}::" }.join("") + @extend_class_name
+        base = @modules.map { |modu| "#{modu}::" }.join("")
+        @class_name =  base + class_name
+        if extend_class_name
+          @extend_class_name = base + extend_class_name
         end
       end