- Cache class_name and extend_class_name
- Don't use `present?` (Because it's slow

## Benchmark

I benchmarked in gitlab repository. https://github.com/gitlabhq/gitlabhq
try 3 times.

### Command

```sh
$ rails_best_practice
```

### Before

- 99.84s
- 100.77s
- 100.19s
- avg: 100.26666666666667s


### After
- 27.07s
- 25.98s
- 25.91s
- avg: 26.319999999999997s