@pocke 2016-07-13
Hmm, travis ci failed.

Travis says the followings.

> Gem::InstallError: activesupport requires Ruby version >= 2.2.2.

ActiveSupport 5 requires ruby 2.2.2 or later. but rubies version on CI aren't larger than 2.2.2.